use same component in both places
