+++ client/src/components/navigation/NavigationBarSeller.js
@@ -18,7 +18,7 @@ const NavigationBarSeller = ({ setSellerView }) => {
           <path d="M7.293.707A1 1 0 0 0 7 1.414V2H2a1 1 0 0 0-1 1v2a1 1 0 0 0 1 1h5v1H2.5a1 1 0 0 0-.8.4L.725 8.7a.5.5 0 0 0 0 .6l.975 1.3a1 1 0 0 0 .8.4H7v5h2v-5h5a1 1 0 0 0 1-1V8a1 1 0 0 0-1-1H9V6h4.5a1 1 0 0 0 .8-.4l.975-1.3a.5.5 0 0 0 0-.6L14.3 2.4a1 1 0 0 0-.8-.4H9v-.586A1 1 0 0 0 7.293.707z" />
         </svg>
       </Nav.Link>
-      <Nav.Link id="add" as={Link} to="/update/29">
+      <Nav.Link id="add" as={Link} to="/add">
         <svg
           xmlns="http://www.w3.org/2000/svg"
           width="16"
