+++ client/src/components/navigation/RoutesSeller.js
@@ -16,6 +16,9 @@ const RoutesSeller = ({ user, logOut, setSellerView }) => (
     <Route exact path="/home">
       <HomePageSeller />
     </Route>
+    <Route path="/add">
+      <AddProducts />
+    </Route>
     <Route path="/update/:id">
       <UpdateProduct />
     </Route>
